requirements.txt

fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0
requests>=2.28.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
aiohttp>=3.8.0
httpx>=0.24.0
modal>=0.63.0
anthropic>=0.25.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0